fix(snack-runtime): consolidate `UrlUtils` and `utils/SnackUrls`
