+++ packages/snack-runtime/src/utils/SnackUrls.ts
@@ -91,3 +91,30 @@ export function createSnackUrlFromHashId(hashId: string): string {
 export function isEphemeralSnackUrl(url = ''): boolean {
   return extractChannelFromSnackUrl(url) !== null;
 }
+
+/**
+ * Parse and extract the Snack or "experience" URL and possible test transport.
+ * Note, this should only be used within Snack itself and is intended to test the socket systems.
+ */
+export function parseExperienceURL(
+  experienceURL: string
+): { channel: string; testTransport: string | null } | null {
+  const matches = experienceURL.match(/(\+|\/sdk\..*-)([^?]*)\??(.*$)/);
+  if (!matches) {
+    return null;
+  }
+  const channel = matches[2];
+
+  let testTransport = null;
+  const queryItems = (matches[3] ?? '').split(/&/g);
+  for (const item of queryItems) {
+    if (item.startsWith('testTransport=')) {
+      testTransport = item.substring(14);
+      break;
+    }
+  }
+  return {
+    channel,
+    testTransport,
+  };
+}
